Check whether comments are entered when the drop button is clicked
Information > Drop button, you can drop the project without writing any comments.
However, I think it will be helpful for future projects when the reason for the drop is clearly stated.
Therefore, when you click the drop button, if you have not entered a comment, a message appears asking you to enter a comment.

Resolves: #108

Signed-off-by: namkyu1999 <lak9348@konkuk.ac.kr>